# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AnalyseFactsRequest": {
        "properties": {
          "contract_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Contract Ids",
            "type": "array"
          },
          "facts": {
            "items": {
              "maxLength": 4000,
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Facts",
            "type": "array"
          },
          "narrative": {
            "default": "",
            "maxLength": 4000,
            "title": "Narrative",
            "type": "string"
          },
          "proceeding_posture": {
            "anyOf": [
              {
                "enum": [
                  "quash",
                  "discharge",
                  "trial",
                  "appeal"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stage of the proceeding: quash/discharge judge whether the record prima facie discloses each element; trial/appeal judge whether the evidence proves it. Omit for the stricter default (proved). No effect unless the engine runs a standard-aware prompt template.",
            "title": "Proceeding Posture"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sections"
          }
        },
        "required": [
          "facts",
          "contract_ids"
        ],
        "title": "AnalyseFactsRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pravrudhi partner API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyse-facts": {
      "post": {
        "operationId": "analyse_facts_ep_api_v1_analyse_facts_post",
        "parameters": [
          {
            "description": "Include second-judge diagnostic fields per element (only for deployments that use two judges) even when not authenticated. Only takes effect when this deployment's own debug_second_judge_fields_enabled is also set -- a caller cannot turn this on for a deployment that hasn't opted in. An authenticated caller (Supabase session or org API key) always gets these fields regardless of this flag or the deployment gate.",
            "in": "query",
            "name": "debug_second_judge",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include second-judge diagnostic fields per element (only for deployments that use two judges) even when not authenticated. Only takes effect when this deployment's own debug_second_judge_fields_enabled is also set -- a caller cannot turn this on for a deployment that hasn't opted in. An authenticated caller (Supabase session or org API key) always gets these fields regardless of this flag or the deployment gate.",
              "title": "Debug Second Judge",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyseFactsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyseFactsResponse"
                }
              }
            },
            "description": "Successful analysis.",
            "headers": {
              "X-RateLimit-Limit": {
                "description": "Calls per minute this API key may make.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Calls left in the current one-minute window after this one.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Reset": {
                "description": "Seconds until the current window ends and the count resets.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "The API key is invalid or revoked."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Over the rate limit; wait Retry-After seconds.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Limit": {
                "description": "Calls per minute this API key may make.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Calls left in the current one-minute window after this one.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Reset": {
                "description": "Seconds until the current window ends and the count resets.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "description": "Outside the service window, judges warming (retry_after_s), service config missing, or the engine at capacity."
          }
        },
        "summary": "Analyse Facts Ep"
      }
    },
    "/api/v1/analyse-facts/jobs": {
      "post": {
        "operationId": "submit_job_ep_api_v1_analyse_facts_jobs_post",
        "parameters": [
          {
            "in": "query",
            "name": "debug_second_judge",
            "required": false,
            "schema": {
              "default": false,
              "title": "Debug Second Judge",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyseFactsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Jobs require a valid API key."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too many unfinished jobs for this key; wait Retry-After seconds.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "description": "Service config missing, or the engine is unavailable."
          }
        },
        "summary": "Submit Job Ep"
      }
    },
    "/api/v1/analyse-facts/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_ep_api_v1_analyse_facts_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Jobs require a valid API key."
          },
          "404": {
            "description": "No such job for this key."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Service config missing."
          }
        },
        "summary": "Get Job Ep"
      }
    },
    "/api/v1/audit": {
      "get": {
        "operationId": "audit_ep_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Ep"
      }
    },
    "/api/v1/orgs": {
      "post": {
        "operationId": "create_org_ep_api_v1_orgs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Org Ep"
      }
    },
    "/api/v1/orgs/{org_id}/keys": {
      "get": {
        "operationId": "list_keys_ep_api_v1_orgs__org_id__keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeysOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Keys Ep"
      },
      "post": {
        "operationId": "create_key_ep_api_v1_orgs__org_id__keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Key Ep"
      }
    },
    "/api/v1/orgs/{org_id}/keys/{key_id}/revoke": {
      "post": {
        "operationId": "revoke_key_ep_api_v1_orgs__org_id__keys__key_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Key Ep"
      }
    },
    "/api/v1/orgs/{org_id}/usage": {
      "get": {
        "operationId": "usage_ep_api_v1_orgs__org_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage Ep"
      }
    },
...
  }
}